Added  cell to prefetch models.
Added a config cell for the Colab instance.  Now it can pre-fetch models.  Folder on Google drive can be specified. If models are found there, they will be symlinked instead of downloaded.  Any models found in folder, but not in download list, will be symlinked to models/custom.
Also added comments to code and titles to cells.
Should just be able to enter settings in first cell, then hit 'run all'.
Well... twice.